+++ src/main/java/RUC/BusquedaRuc1xxx.java
@@ -1,0 +1,10 @@
+package RUC;
+
+import org.apache.poi.xssf.usermodel.XSSFRow;
+import org.openqa.selenium.WebDriver;
+
+public class BusquedaRuc1xxx {
+    public void AsignarRUC1xxx(XSSFRow row, WebDriver driver, String rucString) {
+
+    }
+}
